Rename playlist and home components
