+++ src/Styles/index.scss
@@ -1,7 +1,7 @@
 @import '../App/App.scss';
-@import '../Home/Home.scss';
+@import '../TopArtists/TopArtists.scss';
 @import '../Login/Login.scss';
-@import '../Playlist/Playlist.scss';
+@import '../TopSongs/TopSongs.scss';
 
 * {
   margin: 0;
