+++ src/TopArtists/TopArtists-actions.test.js
@@ -1,7 +1,7 @@
-import * as actions from './Home-actions';
+import * as actions from './TopArtists-actions';
 import { mockAuthCode } from '../Utilities/mockData';
 
-describe('Home actions', ()=> {
+describe('TopArtists actions', ()=> {
   it('should create an action to detect authorization code', ()=>{
     const expectedAction = {
       type: 'AUTH_CODE',
