+++ src/TopArtists/TopArtists.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { authCodeCleaner } from '../Utilities/helpers';
-import { saveAuthCodeAction } from './Home-actions'; 
+import { saveAuthCodeAction } from './TopArtists-actions'; 
 import { Meter } from '../Meter/Meter';
 
-class Home extends React.Component{
+class TopArtists extends React.Component{
 
   constructor(){
     super();
@@ -130,7 +130,7 @@ const mapDispatchToProps = dispatch => {
   };
 };
 
-Home.propTypes = {
+TopArtists.propTypes = {
   authCodeToSagas: PropTypes.func,
   token: PropTypes.string,
   history: PropTypes.object,
@@ -138,4 +138,4 @@ Home.propTypes = {
   userInfo: PropTypes.object,
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(Home);
+export default connect(mapStateToProps, mapDispatchToProps)(TopArtists);
